[chore] rewrite the evaluator
- use a Machine monad and interpret into Eval Monad
- make interpretation to frames explicit
